third_party/WebKit/LayoutTests/external/wpt/fullscreen/model/move-to-fullscreen-iframe-manual-expected.txt

This is a testharness.js-based test.
FAIL Moving fullscreen document's body into a fullscreen iframe assert_equals: iframe's final fullscreen element expected Element node <body></body> but got null
Harness: the test ran to completion.
